include good_job dashboard
