+++ app/views/admin/nav/index.html.erb
@@ -8,6 +8,9 @@
         <%= link_to t('sessions.sessions'), user_sessions_path, class: 'nav__link' %>
       </span>
       <span class="nav__item">
+        <%= link_to t('application.good_job'), good_job_path, class: 'nav__link' %>
+      </span>
+      <span class="nav__item">
         <%= link_to t('sessions.logout'), logout_path, class: 'nav__link' %>
       </span>
     <% end %>
